+++ tests/FunctionalTest/ErrorsTest.php
@@ -6,6 +6,6 @@ class ErrorsTest extends BaseFunctionalTest
 {
     public function testMissingTemplate()
     {
-        $this->assertGenerationError('missing-template-file', 'Template "home.twig" is not defined.');
+        $this->assertGenerationError('missing-layout-file', 'Template "home.twig" is not defined.');
     }
 }